V6: Modify Course Credits to decimal (EF + State-based)

+++ StudentManagement/Models/Course.cs
@@ -4,7 +4,7 @@ namespace StudentManagement.Models
     {
         public int Id { get; set; }
         public string Title { get; set; } = string.Empty;
-        public int Credits { get; set; }
+        public decimal Credits { get; set; }
 
         public int InstructorId { get; set; }
         public Instructor Instructor { get; set; } = null!;
